Add utils file for shared test service
